Refactor dir.c & loc.c
