+++ ghosts.c
@@ -2,6 +2,7 @@
 #define GHOSTS_C_INCLUDED
 
 #include "utils.c"
+#include "loc.c"
 
 // Ghosts type coding
 typedef enum {
